refactored jwt
